requirements.txt
tensorflow==2.12.0
pillow==9.5.0
numpy==1.24.2
